// types/station.ts
export interface Station {
  name: string;
  code: string;
  utterances: string[];
  name_hi?: string;
  name_gu?: string;
  name_ta?: string;
  name_te?: string;
  name_kn?: string;
  name_ml?: string;
  name_mr?: string;
  name_pa?: string;
  name_bn?: string;
  name_or?: string;
  name_as?: string;
  district: string;
  state: string;
  trainCount: string;
  latitude: number;
  longitude: number;
  address: string;
}